to reload after check the agreement...
Hello!

Stripejs helps me a lot, so I want to say Thanks at first.
I want to propose a change to make this code to change to a function.

And I put a code of "takeStripe();" in file of "order-opc.jp". 
after

updatePaymentMethods(json);
if (typeof bindUniform !=='undefined')
bindUniform();

maybe around line 340.

updatePaymentMethods(json);
if (typeof bindUniform !=='undefined')
bindUniform();
takeStripe();


Like this. 
Because, with original code, sometimes for example, after check for agreement of term of service, Ajax not work and it will be Stripe error.

I'm a web designer and I'm not an expert of programming so I don't know this is best solution but now it works for me.